ml/requirements.txt
streamlit==1.36.0
langchain_chroma==0.1.2
langchain_community
langchain-openai
sentence-transformers
protobuf==3.20.*
chromadb==0.5.4
openai
